fix(types): make rollup work with tsconfig's baseUrl
We have configured `baseUrl` in tsconfig.json, which allows absolute imports to be resolved by typescript. However, we didn't configure the same thing in the rollup config, specifically in the `dts` plugin. This led to an issue after the package was published https://github.com/duffelhq/duffel-api-javascript/issues/372. This PR adds a baseUrl configuration to ensure that absolute import can be resolved the same way during the rollup process.