fix(client): fix DuffelError exporting (#440)
Previously DuffelError was exported as both the types and the object. This led to a naming conflict and caused `DuffelError` to not be importable through `@duffel/api`.

To fix that, we will move this class out of `ClientType`, so that it can be imported as both a value and a type.