fix(build): downgrade node-fetch to 2.6.7
node-fetch v3.0 is ESM-only, and upgrading to that means that our library would be unusable for the users who can't afford to migrate to ESM yet. So we will stick with 2.6.7 for the time being. The maintainers promised that bugs will still continue to be handled in the v2 version so we will take their words for it for now.