[STAY-2408] feat: add deposit type to stays payment type (#999)
### Context 

We are starting to offer deposit rates via the Stays API. We are not yet
finished with our implementation, waiting on details from Travelport
around how we can show this data in a structured way. However, to
unblock some customers and gather some data on what these rates look
like, we are going to allow booking these rates.

We aren't adding support for _booking_ these via the dashboard yet, but
once booked both Duffel and other dashboard users will need to view the
booking details, so we need to add this type to add a new price
breakdown component for deposit type rates.

### Implementation

- Add `deposit` to the payment types type.